Bare metal updates.
git-svn-id: http://ellcc.org/svn/ellcc/trunk@3858 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ baremetal/arm/kernel.h
@@ -4,6 +4,8 @@
 #ifndef _kernel_h_
 #define _kernel_h_
 
+#include <stdint.h>
+
 /** Set a system call handler.
  * @param nr The system call number.
  * @param fn The system call handling function.
@@ -19,7 +21,7 @@ int __set_syscall(int nr, void *fn);
  * @param arg1 The first argument ro the entry point.
  * @param arg2 The second argument to the entry point.
  */
-void __new_context(void **savearea, int (*entry)(intptr_t, intptr_t), int mode, void *ret, intptr_t arg1, intptr_t arg2);
+void __new_context(void **savearea, intptr_t (*entry)(intptr_t, intptr_t), int mode, void *ret, intptr_t arg1, intptr_t arg2);
 
 /** Switch to a new context.
  * @param from A place to store the current context.
